Move pseudo util file
